Changed PSC_TW and PS_E_calib brun and evnt arugment types for JANA 0.7.4p1

+++ src/plugins/Calibration/PS_E_calib/JEventProcessor_PS_E_calib.cc
@@ -143,7 +143,7 @@ jerror_t JEventProcessor_PS_E_calib::init(void)
 //------------------
 // brun
 //------------------
-jerror_t JEventProcessor_PS_E_calib::brun(JEventLoop *eventLoop, int runnumber)
+jerror_t JEventProcessor_PS_E_calib::brun(JEventLoop *eventLoop, int32_t runnumber)
 {
 	// This is called whenever the run number changes
 
@@ -169,7 +169,7 @@ jerror_t JEventProcessor_PS_E_calib::brun(JEventLoop *eventLoop, int runnumber)
 //------------------
 // evnt
 //------------------
-jerror_t JEventProcessor_PS_E_calib::evnt(JEventLoop *loop, int eventnumber)
+jerror_t JEventProcessor_PS_E_calib::evnt(JEventLoop *loop, uint64_t eventnumber)
 {
 	// This is called for every event. Use of common resources like writing
 	// to a file or filling a histogram should be mutex protected. Using
